use servlet code
