+++ function-aws-api-proxy-test/build.gradle.kts
@@ -1,7 +1,13 @@
 plugins {
     id("io.micronaut.build.internal.aws-module")
 }
-
+repositories {
+    mavenLocal {
+        mavenContent {
+            snapshotsOnly()
+        }
+    }
+}
 dependencies {
     api(mn.micronaut.http.server)
     api(projects.micronautFunctionAwsApiProxy)
