--- function-aws-api-proxy-test/src/main/java/io/micronaut/function/aws/proxy/test/HttpExchangeToAwsProxyRequestAdapter.java
@@ -1,27 +1,0 @@
-/*
- * Copyright 2017-2025 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.micronaut.function.aws.proxy.test;
-
-import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPEvent;
-import com.sun.net.httpserver.HttpExchange;
-
-/**
- * Creates a {@link APIGatewayV2HTTPEvent} given a {@link HttpExchange}.
- */
-@FunctionalInterface
-public interface HttpExchangeToAwsProxyRequestAdapter {
-    APIGatewayV2HTTPEvent createAwsProxyRequest(HttpExchange httpExchange);
-}
