--- function-aws-api-proxy-test/src/main/java/io/micronaut/function/aws/proxy/test/HttpHandlerApplicationContextAware.java
@@ -1,25 +1,0 @@
-/*
- * Copyright 2017-2025 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.micronaut.function.aws.proxy.test;
-
-import com.sun.net.httpserver.HttpHandler;
-import io.micronaut.context.ApplicationContextProvider;
-
-/**
- * An {@link HttpHandler} that is aware of the Micronaut {@link io.micronaut.context.ApplicationContext}.
- */
-public interface HttpHandlerApplicationContextAware extends HttpHandler, ApplicationContextProvider {
-}
