+++ gradle/libs.versions.toml
@@ -1,8 +1,8 @@
 [versions]
-micronaut = "4.8.3"
+micronaut = "4.8.7"
 micronaut-docs = "2.0.0"
 micronaut-test = "4.7.0"
-groovy = "4.0.22"
+groovy = "4.0.26"
 spock = "2.3-groovy-4.0"
 
 bouncycastle = '1.70'
