+++ test-suite-http-server-tck-function-aws-api-proxy-test/build.gradle.kts
@@ -1,7 +1,13 @@
 plugins {
     id("io.micronaut.build.internal.aws.http-server-tck-module")
 }
-
+repositories {
+    mavenLocal {
+        mavenContent {
+            snapshotsOnly()
+        }
+    }
+}
 dependencies {
     testImplementation(projects.micronautFunctionAwsApiProxyTest)
     testImplementation(projects.micronautFunctionAwsApiProxy)
